% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers-data-wrangling.R
\name{mod_data_wrangling_display_input_variables}
\alias{mod_data_wrangling_display_input_variables}
\title{Display input variables dynamically, according to UI for screening}
\usage{
mod_data_wrangling_display_input_variables(vars, method, ns)
}
\arguments{
\item{vars}{An object holding the data variable names. This is used to display
all the variables in the input variable of the ui.}

\item{method}{User-selected method for data wrangling.}

\item{ns}{A placeholder for Shiny module namespace.}
}
\value{
A set of input variables specific for the user-selected data-wrangling
method.
}
\description{
Display input variables dynamically, according to UI for screening
}
\keyword{internal}
