% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers-ipc-check.R
\name{mod_ipccheck_call_checker}
\alias{mod_ipccheck_call_checker}
\title{Invoke mwana's IPC Acute Malnutrition minimum sample size requirement checker
from within the module server}
\usage{
mod_ipccheck_call_checker(df, cluster, source = character(), area1, area2)
}
\description{
Invoke mwana's IPC Acute Malnutrition minimum sample size requirement checker
from within the module server
}
\keyword{internal}
