% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers-plausibility-check.R
\name{mod_plausibility_call_checker}
\alias{mod_plausibility_call_checker}
\title{Invoke mwana's plausibility checkers dynamically from within module server,
according to user specifications in the UI}
\usage{
mod_plausibility_call_checker(
  df,
  age = NULL,
  sex,
  muac = NULL,
  weight = NULL,
  height = NULL,
  flags,
  area1,
  area2,
  area3,
  .for = c("wfhz", "muac", "mfaz")
)
}
\description{
Invoke mwana's plausibility checkers dynamically from within module server,
according to user specifications in the UI
}
\keyword{internal}
