% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers-plausibility-check.R
\name{mod_plausibility_display_input_variables}
\alias{mod_plausibility_display_input_variables}
\title{Display input variables dynamically, according to UI for screening}
\usage{
mod_plausibility_display_input_variables(vars, method, ns)
}
\description{
Display input variables dynamically, according to UI for screening
}
\keyword{internal}
