% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers-prevalence.R
\name{mod_prevalence_call_combined_prev_estimator}
\alias{mod_prevalence_call_combined_prev_estimator}
\title{Invoke mwana's prevalence functions from within module server according to
user specifications in the UI}
\usage{
mod_prevalence_call_combined_prev_estimator(
  df,
  wts = NULL,
  oedema = NULL,
  area1,
  area2,
  area3
)
}
\description{
Invoke mwana's prevalence functions from within module server according to
user specifications in the UI
}
\keyword{internal}
