% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers-prevalence.R
\name{mod_prevalence_call_prev_estimator_screening}
\alias{mod_prevalence_call_prev_estimator_screening}
\title{Invoke mwana's prevalence functions from within module server according to
user specifications in the UI}
\usage{
mod_prevalence_call_prev_estimator_screening(
  df,
  age,
  muac,
  oedema = NULL,
  area1,
  area2,
  area3
)
}
\arguments{
\item{df, age, muac, oedema, area1, area2, area3}{Input variables collected
from the UI and required to pass to mwana::mw_estimate_prevalence_screening().}
}
\value{
A summary tibble for the descriptive statistics about wasting based
on MUAC, with no confidence intervals.
}
\description{
Invoke mwana's prevalence functions from within module server according to
user specifications in the UI
}
\keyword{internal}
