% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers-prevalence.R
\name{mod_prevalence_display_input_variables}
\alias{mod_prevalence_display_input_variables}
\title{Display input variables dynamically, according to UI for screening}
\usage{
mod_prevalence_display_input_variables(
  vars,
  source,
  indicator_surv,
  has_age,
  ns
)
}
\description{
Display input variables dynamically, according to UI for screening
}
\keyword{internal}
