% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers-prevalence.R
\name{mod_prevalence_neat_output_survey}
\alias{mod_prevalence_neat_output_survey}
\title{Neat prevalence output from survey}
\usage{
mod_prevalence_neat_output_survey(df, .type = c("wfhz", "muac", "combined"))
}
\arguments{
\item{df}{data.frame containing the prevalence results.}

\item{.type}{A choice from which the prevalence is derived.}
}
\value{
A tibble object of the same length and width as df, with column
names and values formatted for clarity and readability.
}
\description{
Neat prevalence output from survey
}
\keyword{internal}
