% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-prevalence.R
\name{module_server_prevalence}
\alias{module_server_prevalence}
\title{Module server for prevalence analysis}
\usage{
module_server_prevalence(id, data)
}
\arguments{
\item{id}{Module ID}

\item{data}{An output and reactive data object from previous tab.}
}
\description{
Module server for prevalence analysis
}
