% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-data-wrangling.R
\name{module_server_wrangling}
\alias{module_server_wrangling}
\title{Module server for data wrangling}
\usage{
module_server_wrangling(id, data)
}
\arguments{
\item{id}{Module ID}

\item{data}{An output and reactive data object from previous tab.}
}
\value{
A reactive data object wrangled and ready to be used in plausibility
checks and prevalence analysis tabs.
}
\description{
Module server for data wrangling
}
\keyword{internal}
