% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-ipc-aman-check.R
\name{module_ui_ipccheck}
\alias{module_ui_ipccheck}
\title{Module UI for IPC Acute Malnutrition sample size requirements check}
\usage{
module_ui_ipccheck(id)
}
\arguments{
\item{id}{Module ID}
}
\description{
Module UI for IPC Acute Malnutrition sample size requirements check
}
\keyword{internal}
