% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-plausibility-check.R
\name{module_ui_plausibility_check}
\alias{module_ui_plausibility_check}
\title{Module UI for plausibility check}
\usage{
module_ui_plausibility_check(id)
}
\arguments{
\item{id}{Module ID}
}
\description{
Module UI for plausibility check
}
\keyword{internal}
