% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-prevalence.R
\name{module_ui_prevalence}
\alias{module_ui_prevalence}
\title{Module UI for prevalence analysis}
\usage{
module_ui_prevalence(id)
}
\arguments{
\item{id}{Module ID}
}
\description{
Module UI for prevalence analysis
}
\keyword{internal}
