% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-data-wrangling.R
\name{module_ui_wrangling}
\alias{module_ui_wrangling}
\title{Module UI for data wrangling}
\usage{
module_ui_wrangling(id)
}
\arguments{
\item{id}{Module ID}
}
\description{
Module UI for data wrangling
}
\keyword{internal}
