% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app.R
\name{run_mwana_app}
\alias{run_mwana_app}
\title{Initialise built-in Shiny application}
\usage{
run_mwana_app(package = "mwanaApp")
}
\arguments{
\item{package}{package name (\code{mwanaApp}).}
}
\description{
Initialise built-in Shiny application
}
\examples{
if (interactive()) run_mwana_app()

}
