% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers-ipc-check.R
\name{mod_ipccheck_call_checker}
\alias{mod_ipccheck_call_checker}
\title{Invoke mwana's IPC Acute Malnutrition minimum sample size requirement checker
from within the module server}
\usage{
mod_ipccheck_call_checker(df, cluster, source = character(), area1, area2)
}
\arguments{
\item{df, cluster, source, area1, area2}{Input variables collected from the UI
and required to pass to mwana::mw_check_ipcamn_ssreq() function.}
}
\value{
A summary tibble containing check results for:
\itemize{
\item n_clusters - the total number of unique clusters or screening or site identifiers;
\item n_obs - the corresponding total number of children in the dataset; and,
\item meet_ipc - whether the IPC AMN requirements were met.
}
}
\description{
Invoke mwana's IPC Acute Malnutrition minimum sample size requirement checker
from within the module server
}
\keyword{internal}
