% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers-plausibility-check.R
\name{mod_plausibility_call_checker}
\alias{mod_plausibility_call_checker}
\title{Invoke mwana's plausibility checkers dynamically from within module server,
according to user specifications in the UI}
\usage{
mod_plausibility_call_checker(
  df,
  age = NULL,
  sex,
  muac = NULL,
  weight = NULL,
  height = NULL,
  flags,
  area1,
  area2,
  area3,
  .for = c("wfhz", "muac", "mfaz")
)
}
\arguments{
\item{df, age, sex, muac, weight, height, flags, area1, area2, area3, .for}{Input
variables collected from the UI and required to pass to
mwana::mw_plausibility_check_wfhz().}
}
\value{
A single-row summary tibble with columns containing the plausibility
check results.
}
\description{
Invoke mwana's plausibility checkers dynamically from within module server,
according to user specifications in the UI
}
\keyword{internal}
