% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers-plausibility-check.R
\name{mod_plausibility_display_input_variables}
\alias{mod_plausibility_display_input_variables}
\title{Display input variables dynamically, according to UI for screening}
\usage{
mod_plausibility_display_input_variables(vars, method, ns)
}
\arguments{
\item{vars}{An object holding the data variable names. This is used to display
all the variables in the input variable of the ui.}

\item{method}{User-selected method for data wrangling.}

\item{ns}{A placeholder for Shiny module namespace.}
}
\description{
Display input variables dynamically, according to UI for screening
}
