% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-helpers.R
\name{mod_prevalence_call_prev_estimator_screening}
\alias{mod_prevalence_call_prev_estimator_screening}
\title{Invoke mwana's prevalence functions from within module server according to
user specifications in the UI}
\usage{
mod_prevalence_call_prev_estimator_screening(
  df,
  muac,
  oedema = NULL,
  area1,
  area2,
  area3
)
}
\description{
Invoke mwana's prevalence functions from within module server according to
user specifications in the UI
}
\keyword{internal}
