% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-data-upload.R,
%   R/module-data-wrangling.R
\name{module_ui_upload}
\alias{module_ui_upload}
\alias{module_server_upload}
\alias{module_ui_wrangling}
\title{Module UI for data upload}
\usage{
module_ui_upload(id)

module_server_upload(id)

module_ui_wrangling(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
A reactive data object to be used in data wrangling tab.
}
\description{
Module UI for data upload

Module server for data upload

Module UI for data wrangling
}
\keyword{internal}
