% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-ipc-aman-check.R
\name{module_server_ipccheck}
\alias{module_server_ipccheck}
\title{Module server for IPC Acute Malnutrition sample size requirements check}
\usage{
module_server_ipccheck(id, data)
}
\arguments{
\item{id}{Module ID}
}
\description{
Module server for IPC Acute Malnutrition sample size requirements check
}
\keyword{internal}
