% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-plausibility-check.R
\name{module_server_plausibility_check}
\alias{module_server_plausibility_check}
\title{Module server for plausibility check}
\usage{
module_server_plausibility_check(id, data)
}
\arguments{
\item{id}{Module ID}
}
\description{
Module server for plausibility check
}
\keyword{internal}
