% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-data-wrangling.R
\name{module_server_wrangling}
\alias{module_server_wrangling}
\title{Module server for data wrangling}
\usage{
module_server_wrangling(id, data)
}
\arguments{
\item{id}{Module ID}
}
\description{
Module server for data wrangling
}
\keyword{internal}
