% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwanaApp-package.R
\docType{package}
\name{mwanaApp-package}
\alias{mwanaApp}
\alias{mwanaApp-package}
\title{mwanaApp: mwana GUI}
\description{
A seamless graphical interface to the mwana R package for data wrangling, plausibility checks, and prevalence estimation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mphimo/mwanaApp}
  \item \url{https://mphimo.github.io/mwanaApp/}
  \item Report bugs at \url{https://github.com/mphimo/mwanaApp/issues}
}

}
\author{
\strong{Maintainer}: Tomás Zaba \email{tomas.zaba@outlook.com} (\href{https://orcid.org/0000-0002-7079-3574}{ORCID}) [copyright holder]

}
\keyword{internal}
